[file_utils] support block files in get_file_length()

+++ base/file_utils.h
@@ -3,6 +3,7 @@
 
 #include <string>
 #include <sys/types.h>
+#include <stdint.h>
 
 //----------------------------------------------------------------
 
@@ -12,7 +13,7 @@ namespace file_utils {
 	void check_file_exists(std::string const &file);
 	int create_block_file(std::string const &path, off_t file_size);
 	int open_block_file(std::string const &path, off_t min_size, bool writeable, bool excl = true);
-	size_t get_file_length(std::string const &file);
+	uint64_t get_file_length(std::string const &file);
 }
 
 //----------------------------------------------------------------
